删除角色逻辑

+++ src/main/java/com/pjb/springbootjwt/zhddkk/controller/SysRoleController.java
@@ -9,11 +9,14 @@ import java.util.List;
 //import org.apache.shiro.authz.annotation.RequiresPermissions;
 import com.baomidou.mybatisplus.enums.SqlLike;
 import com.pjb.springbootjwt.zhddkk.domain.SysRoleMenuDO;
+import com.pjb.springbootjwt.zhddkk.domain.SysUserRoleDO;
 import com.pjb.springbootjwt.zhddkk.service.SysMenuService;
+import com.pjb.springbootjwt.zhddkk.service.SysUserRoleService;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
@@ -35,89 +38,94 @@ import org.slf4j.LoggerFactory;
 @Controller
 @RequestMapping("/zhddkk/sysRole")
 public class SysRoleController extends AdminBaseController {
-
+    
     private static final Logger logger = LoggerFactory.getLogger(SysRoleController.class);
-
+    
     @Autowired
     private SysRoleService sysRoleService;
-
+    
     @Autowired
     private SysRoleMenuService sysRoleMenuService;
-
-	/**
-    * binder.
-	* @param binder binder
-	*/
-	@InitBinder
+    
+    @Autowired
+    private SysUserRoleService sysUserRoleService;
+    
+    /**
+     * binder.
+     * 
+     * @param binder binder
+     */
+    @InitBinder
     public void initBinder(WebDataBinder binder) {
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         dateFormat.setLenient(false);
         binder.registerCustomEditor(Date.class, new CustomDateEditor(dateFormat, true));
     }
-
+    
     /**
-    * 跳转到角色表页面.
-	*/
-	@GetMapping()
-	//@RequiresPermissions("zhddkk:sysRole:sysRole")
-    public String sysRole(){
-	    return "zhddkk/sysRole/sysRole";
-	}
-
+     * 跳转到角色表页面.
+     */
+    @GetMapping()
+    // @RequiresPermissions("zhddkk:sysRole:sysRole")
+    public String sysRole() {
+        return "zhddkk/sysRole/sysRole";
+    }
+    
     /**
      * 获取角色表列表数据.
      */
-	@ResponseBody
-	@GetMapping("/list")
-	//@RequiresPermissions("zhddkk:sysRole:sysRole")
-	public Result<Page<SysRoleDO>> list(SysRoleDO sysRoleDto) {
+    @ResponseBody
+    @GetMapping("/list")
+    // @RequiresPermissions("zhddkk:sysRole:sysRole")
+    public Result<Page<SysRoleDO>> list(SysRoleDO sysRoleDto) {
         Wrapper<SysRoleDO> wrapper = new EntityWrapper<SysRoleDO>(sysRoleDto);
         if (StringUtils.isNotBlank(sysRoleDto.getRoleNameLike())) {
-        	wrapper.like("name", sysRoleDto.getRoleNameLike(), SqlLike.DEFAULT);
-		}
+            wrapper.like("name", sysRoleDto.getRoleNameLike(), SqlLike.DEFAULT);
+        }
         Page<SysRoleDO> page = sysRoleService.selectPage(getPage(SysRoleDO.class), wrapper);
         return Result.ok(page);
-	}
-
+    }
+    
     /**
      * 跳转到角色表添加页面.
      */
-	@GetMapping("/add")
-	//@RequiresPermissions("zhddkk:sysRole:add")
+    @GetMapping("/add")
+    // @RequiresPermissions("zhddkk:sysRole:add")
     public String add(Model model) {
-		SysRoleDO sysRole = new SysRoleDO();
+        SysRoleDO sysRole = new SysRoleDO();
         model.addAttribute("sysRole", sysRole);
-	    return "zhddkk/sysRole/sysRoleForm";
-	}
-
+        return "zhddkk/sysRole/sysRoleForm";
+    }
+    
     /**
      * 跳转到角色表编辑页面.
+     * 
      * @param id 角色表ID
      * @param model 角色表实体
      */
-	@GetMapping("/edit/{id}")
-	//@RequiresPermissions("zhddkk:sysRole:edit")
+    @GetMapping("/edit/{id}")
+    // @RequiresPermissions("zhddkk:sysRole:edit")
     public String edit(@PathVariable("id") Integer id, Model model) {
-		SysRoleDO sysRole = sysRoleService.selectById(id);
-		model.addAttribute("sysRole", sysRole);
-	    return "zhddkk/sysRole/sysRoleForm";
-	}
-	
-	/**
-	 * 保存角色表.
-	 */
-	@ResponseBody
-	@PostMapping("/save")
-	//@RequiresPermissions("zhddkk:sysRole:add")
-	public Result<String> save(SysRoleDO sysRole) {
-		String roleName = sysRole.getName();
-		int roleCount = sysRoleService.selectCount(new EntityWrapper<SysRoleDO>().eq("name", roleName));
-		if (roleCount > 0) {
-			return Result.fail("角色名已存在");
-		}
-		sysRoleService.insert(sysRole);
+        SysRoleDO sysRole = sysRoleService.selectById(id);
+        model.addAttribute("sysRole", sysRole);
+        return "zhddkk/sysRole/sysRoleForm";
+    }
+    
+    /**
+     * 保存角色表.
+     */
+    @ResponseBody
+    @PostMapping("/save")
+    // @RequiresPermissions("zhddkk:sysRole:add")
+    public Result<String> save(SysRoleDO sysRole) {
+        String roleName = sysRole.getName();
+        int roleCount = sysRoleService.selectCount(new EntityWrapper<SysRoleDO>().eq("name", roleName));
+        if (roleCount > 0) {
+            return Result.fail("角色名已存在");
+        }
+        sysRoleService.insert(sysRole);
         List<Integer> menuIdList = sysRole.getMenuIds();
-        if (null != menuIdList && menuIdList.size() > 0){
+        if (null != menuIdList && menuIdList.size() > 0) {
             List<SysRoleMenuDO> insertRoleMenuList = new ArrayList<>();
             for (int menuId : menuIdList) {
                 SysRoleMenuDO sysRoleMenuDO = new SysRoleMenuDO();
@@ -132,18 +140,19 @@ public class SysRoleController extends AdminBaseController {
             sysRoleMenuService.removeByRoleId(sysRole.getId());
         }
         return Result.ok();
-	}
-
-	/**
-	 * 编辑角色表.
-	 */
-	@ResponseBody
-	@RequestMapping("/update")
-	//@RequiresPermissions("zhddkk:sysRole:edit")
-	public Result<String> update(SysRoleDO sysRole) {
-		sysRoleService.updateById(sysRole);
+    }
+    
+    /**
+     * 编辑角色表.
+     */
+    @ResponseBody
+    @RequestMapping("/update")
+    @Transactional
+    // @RequiresPermissions("zhddkk:sysRole:edit")
+    public Result<String> update(SysRoleDO sysRole) {
+        sysRoleService.updateById(sysRole);
         List<Integer> menuIdList = sysRole.getMenuIds();
-        if (null != menuIdList && menuIdList.size() > 0){
+        if (null != menuIdList && menuIdList.size() > 0) {
             List<SysRoleMenuDO> insertRoleMenuList = new ArrayList<>();
             for (int menuId : menuIdList) {
                 SysRoleMenuDO sysRoleMenuDO = new SysRoleMenuDO();
@@ -152,34 +161,43 @@ public class SysRoleController extends AdminBaseController {
                 sysRoleMenuDO.setMenuId(menuId);
                 insertRoleMenuList.add(sysRoleMenuDO);
             }
-
+            
             sysRoleMenuService.removeByRoleId(sysRole.getId());
             sysRoleMenuService.batchSave(insertRoleMenuList);
-        } else {
+        }
+        else {
             sysRoleMenuService.removeByRoleId(sysRole.getId());
         }
-		return Result.ok();
-	}
-	
-	/**
-	 * 删除角色表.
-	 */
-	@PostMapping("/remove")
-	@ResponseBody
-	//@RequiresPermissions("zhddkk:sysRole:remove")
-	public Result<String> remove(Integer id) {
-		sysRoleService.deleteById(id);
         return Result.ok();
-	}
-	
-	/**
-	 * 批量删除角色表.
-	 */
-	@PostMapping("/batchRemove")
-	@ResponseBody
-	//@RequiresPermissions("zhddkk:sysRole:batchRemove")
-	public Result<String> remove(@RequestParam("ids[]") Integer[] ids) {
-		sysRoleService.deleteBatchIds(Arrays.asList(ids));
-		return Result.ok();
-	}
+    }
+    
+    /**
+     * 删除角色表.
+     */
+    @PostMapping("/remove")
+    @ResponseBody
+    @Transactional
+    // @RequiresPermissions("zhddkk:sysRole:remove")
+    public Result<String> remove(Integer id) {
+        SysRoleDO sysRoleDO = sysRoleService.selectById(id);
+        if (sysRoleDO != null) {
+            sysRoleService.deleteById(id);
+            sysUserRoleService.delete(new EntityWrapper<SysUserRoleDO>().eq("role_id", id));
+            sysRoleMenuService.delete(new EntityWrapper<SysRoleMenuDO>().eq("role_id", id));
+            return Result.ok();
+        }
+        
+        return Result.fail();
+    }
+    
+    /**
+     * 批量删除角色表.
+     */
+    @PostMapping("/batchRemove")
+    @ResponseBody
+    // @RequiresPermissions("zhddkk:sysRole:batchRemove")
+    public Result<String> remove(@RequestParam("ids[]") Integer[] ids) {
+        sysRoleService.deleteBatchIds(Arrays.asList(ids));
+        return Result.ok();
+    }
 }
